lazy load LinkBox and psl
